+++ index.html
@@ -7,6 +7,6 @@
     <title>Document</title>
   </head>
   <body>
-    <script src="./calculator-class.js"></script>
+    <script src="./js/calculator-class.js"></script>
   </body>
 </html>
