+++ js/tests/reverse-string.test.js
@@ -1,4 +1,4 @@
-const reverseString = require('./reverse-string');
+const reverseString = require('../reverse-string');
 
 test('reverse string', () => {
   expect(reverseString('open')).toBe('nepo');
